% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gab2sq}
\alias{gab2sq}
\title{Converts an auxiliary block design to a
square array design}
\usage{
gab2sq(m, nc, abd)
}
\arguments{
\item{m}{Number of blocks m}

\item{nc}{Number of rows}

\item{abd}{Auxiliary block design}
}
\value{
sqmat square array design
}
\description{
Converts an auxiliary block design to a
square array design
}
