% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gac}
\alias{gac}
\title{Metrics from Ac}
\usage{
gac(m, nc, Ac)
}
\arguments{
\item{m}{number of treatments}

\item{nc}{number of controls}

\item{Ac}{Ac}
}
\value{
Ac,Acc,Act,Att
}
\description{
Metrics from Ac
}
