% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gc2s}
\alias{gc2s}
\title{Converts coordinates to spacing for cyclic designs}
\usage{
gc2s(m, nc, gc)
}
\arguments{
\item{m}{Number of treatments}

\item{nc}{Number of rows/controls}

\item{gc}{Coordinates -  vector}
}
\value{
gs Spacing
}
\description{
Converts coordinates to spacing for cyclic designs
}
