% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gcount}
\alias{gcount}
\title{isometric classes}
\usage{
gcount(g1, mt, gtot)
}
\arguments{
\item{g1}{number of classes}

\item{mt}{number to print per class}

\item{gtot}{num X 2nc+4 matrix ordered
with cols gc,gs,Ac,Acc,Act,Att}
}
\value{
matrix with top mt from gtot for each class
}
\description{
isometric classes
}
