% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gs2c}
\alias{gs2c}
\title{Converts spacing to coordinates for cyclic designs}
\usage{
gs2c(m, nc, i1, space)
}
\arguments{
\item{m}{Number of treatments}

\item{nc}{Number of rows/controls}

\item{i1}{Starting coordinate in first row}

\item{space}{Spacing - vector}
}
\value{
gc    Coordinates
}
\description{
Converts spacing to coordinates for cyclic designs
}
