% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{gvabd}
\alias{gvabd}
\title{Calculates treatment information matrix C and metric Aabd
for an auxiliary block design}
\usage{
gvabd(m, nc, abmat)
}
\arguments{
\item{m}{Number of treatments}

\item{nc}{Number of rows}

\item{abmat}{Auxiliary block design r x t}
}
\value{
.     Metric Aabd  0 if matrix C is not connected
}
\description{
Calculates treatment information matrix C and metric Aabd
for an auxiliary block design
}
